Move batched dot again
