+++ src/pykeen/nn/modules.py
@@ -32,7 +32,6 @@ from typing_extensions import Self
 from . import functional as pkf
 from . import init
 from .algebra import quaterion_multiplication_table
-from .functional import batched_dot
 from .sim import KG2ESimilarity, kg2e_similarity_resolver
 from .utils import apply_optional_bn
 from ..metrics.utils import ValueRange
@@ -50,6 +49,7 @@ from ..typing import (
 from ..utils import (
     add_cudnn_error_hint,
     at_least_eps,
+    batched_dot,
     clamp_norm,
     einsum,
     ensure_complex,
